Bugfix, apperently the earth isn't flat.
MAX_DISTANCE was big enough to include north pole which apperently is
the location given if gps not found.

In getLocationDifference I replaced our homebrewn distance calculator with the one in
Location. How long a degree latitude resp. longitude is differcs from locations so ours
was not accurate. This was done mainly to allow better max distances.

Sorry for the formating.